Simplify and consolidate
