+++ xknx/dpt/dpt_252.py
@@ -2,21 +2,18 @@
 
 from __future__ import annotations
 
-from collections.abc import Mapping
 from dataclasses import dataclass, field
-from typing import Any
 
-from .dpt import DPTComplex, DPTComplexData
+from .dpt import DPTComplex
 from .dpt_1 import Step
-from .dpt_3 import (
-    RANGE_STEP_CODE,
-    ControlDimming,
+from .dpt_3 import ControlDimming
+from .helpers.metadata import RANGE_STEP_CODE
+from .helpers.relative_control import (
+    _RelativeControlDimming,
     pack_control_dimming,
     unpack_control_dimming,
 )
 from .payload import DPTArray, DPTBinary
-
-_FIELDS = ("red", "green", "blue", "white")
 
 
 @dataclass
@@ -34,7 +31,7 @@ class _RelativeControlRGBWDictSchemaFields:
 
 
 @dataclass(slots=True)
-class RelativeControlRGBW(DPTComplexData):
+class RelativeControlRGBW(_RelativeControlDimming):
     """
     Representation of a relative RGBW color control.
 
@@ -47,43 +44,6 @@ class RelativeControlRGBW(DPTComplexData):
     white: ControlDimming | None = None
 
     _dict_schema_fields_class = _RelativeControlRGBWDictSchemaFields
-
-    @classmethod
-    def from_dict(cls, data: Mapping[str, Any]) -> RelativeControlRGBW:
-        """Init from a dictionary."""
-        result: dict[str, ControlDimming | None] = {}
-        for key in _FIELDS:
-            control = data.get(f"{key}_control")
-            step_code = data.get(f"{key}_step_code")
-            if control is not None and step_code is not None:
-                try:
-                    result[key] = ControlDimming(
-                        control=Step.parse(control), step_code=int(step_code)
-                    )
-                except (TypeError, ValueError) as err:
-                    raise ValueError(f"Invalid value for {key}: {err}") from err
-            elif control is not None or step_code is not None:
-                raise ValueError(
-                    f"Provide both {key}_control and {key}_step_code, or neither"
-                )
-            else:
-                result[key] = None
-        return cls(**result)
-
-    def as_dict(self) -> dict[str, int | str | None]:
-        """Create a JSON serializable dictionary."""
-        result: dict[str, int | str | None] = {}
-        for key, value in (
-            ("red", self.red),
-            ("green", self.green),
-            ("blue", self.blue),
-            ("white", self.white),
-        ):
-            result[f"{key}_control"] = (
-                value.control.name.lower() if value is not None else None
-            )
-            result[f"{key}_step_code"] = value.step_code if value is not None else None
-        return result
 
 
 class DPTRelativeControlRGBW(DPTComplex[RelativeControlRGBW]):
