+++ xknx/dpt/helpers/__init__.py
@@ -1,0 +1,1 @@
+"""Internal helpers shared across DPT implementations."""
